Fix: "Values of identifierUris property must use a verified domain of the organization or its subdomain"
Full error:
Changing "aks-getting-started-sp" to a valid URI of "http://aks-getting-started-sp", which is the required format used for service principal names
Values of identifierUris property must use a verified domain of the organization or its subdomain: 'http://aks-getting-started-sp'

Replaced with:
WARNING: 'name' property in the output is deprecated and will be removed in the future. Use 'appId' instead.